refactor: use singular form of `controller`
