+++ internal/controller/admin.go
@@ -1,4 +1,4 @@
-package controllers
+package controller
 
 import (
 	"time"
