+++ internal/controller/stage.go
@@ -1,4 +1,4 @@
-package controllers
+package controller
 
 import (
 	"github.com/gofiber/fiber/v2"
