+++ internal/controller/v2/formula.go
@@ -1,4 +1,4 @@
-package shims
+package v2
 
 import (
 	"github.com/gofiber/fiber/v2"
