+++ internal/controller/v2/item.go
@@ -1,4 +1,4 @@
-package shims
+package v2
 
 import (
 	"time"
