+++ internal/controller/v2/private.go
@@ -1,4 +1,4 @@
-package shims
+package v2
 
 import (
 	"time"
