+++ internal/controller/v2/report.go
@@ -1,4 +1,4 @@
-package shims
+package v2
 
 import (
 	"encoding/json"
