+++ internal/controller/v2/site_stats.go
@@ -1,4 +1,4 @@
-package shims
+package v2
 
 import (
 	"time"
