+++ internal/controller/v2/zone.go
@@ -1,4 +1,4 @@
-package shims
+package v2
 
 import (
 	"time"
